Refactored io module
