+++ vent/io/iobase.py
@@ -1,32 +1,34 @@
-import pigpio
-import abc
+from pigpio import pi as PigPi
+from abc import ABC, abstractmethod
 import numpy as np
 from collections import OrderedDict
-from time import sleep
+
 
 class Ventilator:
     '''
     Class representation of the physical ventilator device
     '''
     def __init__(self):
-        self.pig = pigpio.pi()
-	
+        #self.pig = PigPi()
+        pass
+        
     def __del__(self):
-        self.pig.stop()
-
+        #self.pig.stop()
+        pass
+        
     def test(self,sensor):
-        return result
+        pass
 
     def calibrate(self):
-        return
+        pass
 
-
-class Device(abc.ABC):
+class IODeviceBase(ABC):
     '''
-    Abstract class for pigpio handles
+    Abstract base Class for pigpio handles (or whatever other GPIO library
+    we end up using)
     '''
     def __init__(self,pig):
-        self._pig = pig if pig is not None else pigpio.pi()
+        self._pig = pig if pig is not None else PigPi()
 
     def __del__(self):
         self.close()
@@ -34,56 +36,62 @@ class Device(abc.ABC):
     @property
     def handle(self):
         return self._handle
+        
+    def open_i2c(self,i2c_bus,i2c_address):
+        self._handle = self._pig.i2c_open(i2c_bus,i2c_address)
+        
+    def open_spi(self,channel,baudrate):
+        self._handle = self._pig.spi_open(channel,baudrate)
 
     def close(self):
         self._pig.close(self._handle)
 
-class i2cDevice(Device):
+class i2cDevice(IODeviceBase):
     '''
     A class wrapper for pigpio I2C handles 
     '''
     def __init__(self, i2c_address, i2c_bus, pig=None):
         super().__init__(pig)
-        self._handle    = self._pig.i2c_open(i2c_bus,i2c_address)
+        self.open_i2c(i2c_bus,i2c_address)
 
     def read_device(self,num_bytes):
         '''
         Read a specified number of bytes directly from the the device 
         without changing the register. Does NOT perform LE/BE conversion
         '''
-        return self._pig.i2c_read_device(self._handle,num_bytes)
+        return self._pig.i2c_read_device(self.handle,num_bytes)
 
     def write_device(self,data):
         '''
         Write bytes to the device without specifying register.
         Does NOT perform LE/BE conversion
         '''
-        self._pig.i2c_write_device(self._handle,data)
+        self._pig.i2c_write_device(self.handle,data)
 
     def read_word(self):
         '''
         Read two words directly from the the device 
         without changing the register. 
         '''
-        return self._be16_to_le(self._pig.i2c_read_word(self._handle))
+        return self._be16_to_le(self._pig.i2c_read_word(self.handle))
 
     def write_word(self,word):
         '''
         Write two bytes to the device without specifying register.
         '''
-        self._pig.i2c_write_word(self._handle,self._le16_to_be(word))
+        self._pig.i2c_write_word(self.handle,self._le16_to_be(word))
 
     def read_register(self,register):
         '''
         Read 2 bytes from the specified register(denoted by a single byte)
         '''
-        return self._be16_to_le(self._pig.i2c_read_word_data(self._handle,register))
+        return self._be16_to_le(self._pig.i2c_read_word_data(self.handle,register))
 
     def write_register(self,register,word):
         '''
         Write 2 bytes to the specified register (denoted by a single byte)
         '''
-        self._pig.i2c_write_word_data(self._handle,register,self._le16_to_be(word))
+        self._pig.i2c_write_word_data(self.handle,register,self._le16_to_be(word))
 
     def _be16_to_le(self,word):
         '''BigEndian to LittleEndian conversion for signed 2 Byte integers (2 complement).'''
@@ -174,95 +182,17 @@ class i2cDevice(Device):
                 '''Extracts setting from passed 16-bit config & returns integer representation'''
                 return ( cfg & (self._MASK<<self._OFFSET) )>>self._OFFSET
 
-class spiDevice(Device):
+class spiDevice(IODeviceBase):
     '''
     A class wrapper for pigpio SPI handles. Not implemented.
     '''
-    def __init(self,pig,spi_channel,baudrate):
+    def __init(self,pig,channel,baudrate):
         super().__init__(pig)
-        self._handle = self._pig.spi_open(spi_channel,baudrate)
-        	
-class ads1115(i2cDevice):  
-    '''
-    Class for the ADS1115 16 bit, 4 Channel Analog to Digital Converter.
-    Datasheet:
-        http://www.ti.com/lit/ds/symlink/ads1114.pdf?ts=1587872241912
-    '''
-    
-    '''Default Values'''
-    #Default ventidude config:   '0b1100001111100011' / 0xC3E3 / 50147
-    #Default config on power-up: '0b1000010110000101' / 0x8583 / 34179
-    _DEFAULT_ADDRESS        = 0x48
-    _DEFAULT_VALUES         = {'MUX':0, 'PGA':4.096, 'MODE':'SINGLE', 'DR':860}
-    
-    '''Address Pointer Register (write-only)'''
-    _POINTER_FIELDS = ( 'P' )
-    _POINTER_VALUES = ( ('CONVERSION', 'CONFIG', 'LO_THRESH', 'HIGH_THRESH'), )
-    
-    '''Config Register (R/W) '''
-    _CONFIG_FIELDS = ('OS','MUX','PGA','MODE','DR','COMP_MODE','COMP_POL','COMP_LAT','COMP_QUE')   
-    _CONFIG_VALUES  = ( ( 'NO_EFFECT', 'START_CONVERSION' ),
-                        ( (0, 1), (0, 3), (1, 3), (2, 3), 0, 1, 2, 3 ),
-                        ( 6.144, 4.096, 2.048, 1.024, 0.512, 0.256, 0.256, 0.256 ),
-                        ( 'CONTINUOUS', 'SINGLE' ),
-                        ( 8, 16, 32, 64, 128, 250, 475, 860 ),
-                        ( 'TRADIONAL', 'WINDOW' ),
-                        ( 'ACTIVE_LOW', 'ACTIVE_HIGH' ),
-                        ( 'NONLATCHING', 'LATCHING' ),
-                        ( 1, 2, 3, 'DISABLE' ) )
-                        
-    '''
-    Note: The Conversion Register is read-only and contains a 16bit representation of 
-    the requested value (provided the conversion is ready).
-    The Lo-thresh & Hi-thresh Registers are not used in this application.
-    However, their function and usage are described in the datasheet. 
-    '''
-    
-    def __init__(self, address=_DEFAULT_ADDRESS, i2c_bus=1, pig=None,):
-        super().__init__(address,i2c_bus,pig)
-        '''Define registers. Pointer register is write only, config is R/W.'''
-        self.pointer    = self.Register(self._POINTER_FIELDS,self._POINTER_VALUES)
-        self.config     = self.Register(self._CONFIG_FIELDS,self._CONFIG_VALUES)
-        '''Set initial value of _LAST_CFG to what is actually on the ADS'''
-        self._LAST_CFG  = self.read_register(self.pointer.P.pack('CONFIG'))
-        '''Pack default settings into _CFG, don't bother to write to ADC yet'''
-        self._CFG       = self.config.pack( cfg     = self._LAST_CFG,
-                                            MUX     = self._DEFAULT_VALUES['MUX'],
-                                            PGA     = self._DEFAULT_VALUES['PGA'],
-                                            MODE    = self._DEFAULT_VALUES['MODE'],
-                                            DR      = self._DEFAULT_VALUES['DR'] )
-                                            
-    def read(self,channel=None,gain=None,mode=None,data_rate=None):
-        '''Performs a raw_read and converts the result to voltage '''
-        return self.raw_read(channel=channel,gain=gain,mode=mode,data_rate=data_rate)*self.config.PGA.unpack(self._CFG) / 32767
-    
-    def raw_read(self,channel=None,gain=None,mode=None,data_rate=None):
-        '''
-        Packs any new values passed as arguments into a new cfg. 
-        If new cfg differs from the last, or if single-shot mode is specified,
-        write new cfg to config register and wait for conversion.
-        Otherwise, or after the above has been done, read the conversion value.
-        '''
-        self._CFG = self.config.pack(self._CFG,MUX=channel,PGA=gain,MODE=mode,DR=data_rate)
-        mode = self.config.MODE.unpack(self._CFG)
-        if self._CFG != self._LAST_CFG or mode == 'SINGLE':
-            self.write_register(self.pointer.P.pack('CONFIG'), self._CFG)
-            sleep(1/self.config.DR.unpack(self._CFG))
-            while not ( self.ready() or  mode == 'CONTINUOUS' ):
-                sleep(1/self.config.DR.unpack(self._CFG)/10)
-                #pass       # not sure which is better here
-        self._LAST_CFG = self._CFG
-        return self.read_register(self.pointer.P.pack('CONVERSION'))
-    
-    def ready(self):
-        '''Return status of ADC conversion.'''
-        # OS = 0: Device is currently performing a conversion
-        # OS = 1: Device is not currently performing a conversion
-        return self.read_register(self.pointer.P.pack('CONFIG')) & (1 << self.config.OS.offset())
-        
+        self.open_spi(hannel,baudrate)
+
 class Sensor(abc.ABC):
     '''
-    Class describing generalized sensors
+    Abstract base Class describing generalized sensors
     '''
     def __init__(self):
         _DATA   = np.zeros(128)
@@ -276,63 +206,19 @@ class Sensor(abc.ABC):
     def n(self):
         return self._N
         
-    @abc.abstractmethod
+    @abstractmethod
     def read(self):
         self._N +=1
     
     def reset(self):
         self._DATA  = np.zeros(128)
         self._N     = 0
-
-    
-class SFM3200(Sensor,i2cDevice):
-    '''
-    Datasheet:
-        https://www.sensirion.com/fileadmin/user_upload/customers/sensirion/Dokumente/5_Mass_Flow_Meters/Datasheets/Sensirion_Mass_Flow_Meters_SFM3200_Datasheet.pdf
-    '''
-    _DEFAULT_ADDRESS     = 0x40 # SFM3x00 Flow sensor address
-    _FLOW_OFFSET         = 32768
-    _FLOW_SCALE_FACTOR   = 120
-
-    def __init__(self,pig,address=_DEFAULT_ADDRESS):
-        i2cDevice.__init__(pig,address)
-        Sensor.__init__()
-        self.reset()
-        self._start()
-
-    def read(self):
-        return self._convert(self._raw_read())
-
-    def _raw_read(self,pig):
-        '''
-        Performs a read on the sensor, converts recieved bytearray, 
-        discards the last two bytes (crc values - could implement in future),
-        and returns a signed int converted from the big endian two 
-        complement that remains.
-        '''
-        return int.from_bytes(self.read_device(4)[2],'big',signed=True)
-
-    def reset(self):
-        super.reset()
-        self.write_word(0x2000)  # soft reset
-        time.sleep(.08)                     # 80 ms soft reset time 
-
-    def _start(self):
-        self.write_word(0x1000)  # start measurement
-        time.sleep(.1)                     # start-up time is 100 ms.
-
-    def _convert(self,value):
-        # Convert raw int to a flow reading represaented by a floating-point
-        # number with units slm.
-        # Source: https://www.sensirion.com/fileadmin/user_upload/customers/sensirion/Dokumente/5_Mass_Flow_Meters/Application_Notes/Sensirion_Mass_Flo_Meters_SFM3xxx_I2C_Functional_Description.pdf
-        return (value-self._FLOW_OFFSET)/self._FLOW_SCALE_FACTOR
-	
+        
 class AnalogSensor(Sensor):
     '''
     General class describing an anlog sensor attached to an ADC.
     If instantiated without a subclass, represents a voltmeter with range 0-1.0.
     '''
-
     def __init__(   self, adc, channel,
                     calibration=(0,5,),
                     gain=None, data_rate=None, mode=None ):
@@ -375,41 +261,21 @@ class AnalogSensor(Sensor):
         Scales raw voltage into the range 0 - 1 
         '''
         return raw - calibration[0]/(calibration[1]-calibration[0])
-            
-class P4vMini(Sensor):
-    '''
-        Analog gauge pressure sensor with range of 0 - 20" h20.
-        Datasheet calibration low is 0.25V and high is 4.0V
-    '''
-    _default_offset_voltage = 0.25
-    _default_output_span    = 4
-    ''' Maximum reading in cm h20: 2.54 cm/in * 20" h20 * 1.0 normalized output '''
-    _conversion_factor      = 2.54*20 
 
-    def __init__(   self, adc, channel,
-                    calibration=(0.25,4.0),
-                    gain=1, data_rate=860, mode=None ):
-        super().__init(self, adc, channel, calibration, gain, data_rate, mode)
 
-    def _convert(self,raw):
-        return self._conversion_factor*self.read()
+#TODO
+class Pin(IODeviceBase)
+	def __init__(self):
+		pass
 
-# ~ class Pin
+class InputPin(Pin)
+	def __init__(self):
+		pass
 
-# ~ class InputPin
+class OutputPin(Pin)
+	def __init__(self):
+		pass
 
-# ~ class OutputPin
-
-# ~ class PWMOutput
-
-# ~ class Valve(abc.ABC):
-	
-# ~ class SolenoidValve(Valve,OutputPin):
-	
-# ~ class ProportionalValve(Valve,PWMOuput): # or DAC out if it comes to that
-
-# ~ class OxygenSensor(AnalogSensor):
-	
-# ~ class HumiditySensor(Sensor):
-	
-# ~ class TemperatureSensor(Sensor):
+class PWMOutput(OutputPin)
+	def __init__(self):
+		pass
